Add spec/alternate handling of entries without recognized agencies
Previously, if an entry was associated with multiple unknown agencies, the entry would be listed a single time with a concatenated set of its agency names (e.g. ‘Agency 1 and Agency 2’.  Now, the entry will be listed in the JSON once for each unknown agency.

Add a spec around ensuring agencies without agency names are rendered.