Update: Improve ES index performance time by ~30%-50%:
Lots of N+1 querying was occurring when indexing.  When you chain additional scopes or pluck within the context of an attribute within the fast JSON api gem, it does not use pre-loaded tables from .includes clauses.

Change this behavior so we’re minimizing scope-chaining and plucking from within the serializer itself and perform this logic in Ruby.

Also added additional tables to the .includes clause in the bulk index task.