Bugfix: Add additional spec/handling around citing document number.
This endpoint was blowing up when using ES as search infrastructure:

/api/v1/documents.json?conditions[citing_document_numbers]=2020-02511&infrastructure=es&metadata_only=1

At a cursory glance, one would expect us to simply be able to return entries.map(&:id) in the proc.  However, in the way the existing search infrastructure is designed, multiple filter objects are shoveled into `search.filters`, hence if the search conditions were citing_document_numbers: [‘abc’, ‘def’] two filters (NOT one) would be created.  This could be re-architected at some point, but it would require a fundamental change to several method calls touched by basically all the search infrastructure and the solution here does what we need it to do.