Update: More closely mirror Thinking Sphinx delta logic when building entry changes:
See the equivalent delta callback in Thinking Sphinx:
https://github.com/pat/thinking-sphinx/blob/1adcf8159fdbcb96a70fb01f5d8a15c2070090e6/lib/thinking_sphinx/active_record/callbacks/delta_callbacks.rb#L50-L52

The EntryChange counts are sometimes slightly higher than the entry deltas.  In a production setting, we’ll probably want to retain this to be conservative, but as we’re continuing to test in staging, it may be useful to see if the counts will align more closely.