Update: Aggressively mark entries as delta in importer.
It appears that both in Sphinx and ES, when agency name assignments occur, entries are not being correctly associated with agencies since the records are not always being marked as delta.  The intent of this change is to mark all records as delta when reimporting such that the indexer (sphinx, es, etc.) reindexes all the entries.