Bugfix: Coerce graphic content types originally stored as GIFs in S3.
In production, it appears graphics stored with these kinds of graphic content types exist on S3, but not as pbm, ppm, or png, but as gifs.  Downstream from paperclip, we're going to coerce these urls accordingly since paperclip is now deprecated and in particular the reprocess command is broken.

'application/x-pbm',
'application/x-ppm',
'image/gif'