Bugfix: Ensure mods synchronously download before reprocessing the issue so the diff is saved.
We were running into an issue where things seem to reprocess successfully, but the diff was not visible in the UI because the #download_mods command and the #reprocess_issue command were both being processed async.  It’s necessary to ensure the mods are downloaded first so the diff is visible.