Update: Remove medium style from graphic object.
Originally, we had planned on regenerating medium-sized images for both graphics and gpo graphics, but to minimize complexity we decided to focus on regenerating strictly gpo graphics.  If a ‘medium’ paperclip style is defined, by default it will end up in the API representation (even if the image does not exist), which will ultimately cause confusion since when a document is reprocessed, FR Web is set to prefer “medium” images and as such, the XSLT processing will be linking up medium.gif image references that don’t exist.