Fix: Handle entry additions triggered via MODS update.
In some cases a MODS change can trigger the removal of a document. If that document is later re-added to the MODS, it will be reimported, but it will be missing the full-text and hence the Web UI will not show the full text and only the PDF will be available.

When a MODS change results in the addition of a new document, trigger a recompile of the full text HTML.

See #199