Fix: Add pagination for recently modified docs
Prior to this change, we had set a hard stop on the number of results that could be returned via the nightly recently modified documents job.  Most of the time this was fine, but occasionally result sets would exceed the hard cap of 250 results.

This fix adds pagination so we no longer encounter this limit.