Update: Comment closing robustness
Immediately remove comment url from documents we anticipate are closing based on the regulations.gov metadata we've previously imported, but also enqueue a job that re-adds the comment url if the regs.gov API indicates it's still open for comment.  The regulations.gov jobs are throttled (and so it takes time to process through them)--we want to present a better UI by not displaying the document as being open for comment once midnight has elapsed so late-night users don't receive an error message about the document not being open for comment when submission is attempted.

See #209