Fix: Searches with prepositions like "of"
If a full text PI document included the phrase "university of california", if you searched for "university of california" in the search box, no results woudl be returned.  It appears this was related to the docket_id mapping.  The docket_id was included as a part of the simple query string and its inclusion with the rest of the title, full_text, and agency_name fields seems to have caused the search to behave in an unexpected manner.

This change indexes the docket_id in the same way as the title.