Fix: Googleauth authentication
Upgrading to a recent version of Carrierwave required an upgrade to addressable > 2.6 (it was previously pinned to 2.4).  Running addressable on 2.8 (the latest at the time) caused the authentication with the googleauth gem to fail--see code snippet below.  The code snippet below was used as a compatibility test for google auth.  i.e. It fails with 2.8, but succeeds with 2.6 and so we're pinning the gem version there.

```
include PageViewCountUtils
page_view_type = PageViewType.first
GoogleAnalytics::PageViews.new.counts(
  default_args.merge(args)
)
GoogleAnalytics::PageViews.new.counts(
  {
    dimension_filters: [
      {
        filters: [
          {
            dimensionName: "ga:pagePath",
            operator: "REGEXP",
            expressions: page_view_type.filter_expressions
          }
        ]
      }
    ],
  }
)
```