Update: Don't update permissions when file blank
Don't attempt to change the ACL when no image file is associated with the record lest the S3 client throw an error.  This was being triggered by the DailyIssueImageUsageBuilder in staging since it attempts to make images public.  This is likely mostly a staging-related concern, but adding handling here for robustness.