Fix: Agency reindexing race condition
After agency name updates, jobs are enqueued to reindex elasticsearch agency_ids and recompile table of contents pages.  We were encountering a race condition when reindexing because the Sidekiq job was looking up the DB record faster than it could be committed, causing some records not to be indexed.

The solution was to use an after_commit callback to ensure the record is
committed before processing enqueuing any job processing.