Fix: FR index failures/admin emails at year-end.
Previously, we were using the date the importer was run to calculate the available years for the FR Index.  This fails at year-end because the importer could be run in a year different than the issue's year.

Downstream, this looked to have been causing the FR index not to update since ActiveRecord::RecordNotFound are intentionally raised if you are attempting to run the FrIndexPresenter for an incorrect year.  Additionally, this was likely causing the admin email to not go out since the import :full rake task was failing at the very end.

See #232.