Fix: Errant marking of failures as "complete"
We handle a number of import-related errors in the #handle_failure method, where we set the reprocessed issue's status to "failed".  At the end of the process we seem to have been marking the issue as complete anyway.  Ensure we keep a "failed" status of a reprocess if that has occurred.