Fix: Long-running async back issue reprocessing
The IssueReprocessor manually invokes several rake tasks and this appears to have the unintended consequence of somehow polluting the sidekiq process such that subsequent runs of the IssueReprocessor take increasingly longer amounts of time to execute.  In lieu of taking 2 minutes to reprocess, they begin to take up to 45 minutes per report.