Fix: Tune batch size so mysql doesn't do full entry table scans
During the pre-joining of the scope used for ES indexing, the mysql query engine falls back to a full table scan of the entries table when calculating corrections instead of using the db index if the number of records in the IN clause exceeds ~325.

This is especially apparent when recently modified documents updates occur at the 55th minute on the hour.