Fix: Apply normalization search pipeline at runtime for hybrid search
The Opensearch documentation states that a "search_pipeline" can be specified via the search_pipeline URL parameter, but our ES client library doesn't seem to like that.  A workaround is to specify the pipeline at search runtime as a part of the search request body in lieu of registering the pipeline in advance.  This seems to be a relatively lightweight thing currently, but it seems like there might be some benefit in registering the pipeline once and re-using it in lieu of having to POST the configuration over with every search request.